feat: update to tables for bods data
