+++ README.md
@@ -2,13 +2,13 @@
 
 [![CircleCI](https://circleci.com/gh/fares-data-build-tool/fdbt-reference-data-service.svg?style=svg)](https://circleci.com/gh/fares-data-build-tool/fdbt-reference-data-service)
 
-Code for the lambdas responible for retrieving the reference data for the Fares Data Build Tool and uploading it to a MySQL Database. This includes NaPTAN, NOC and TNDS data.
+Code for the lambdas responible for retrieving the reference data for the Fares Data Build Tool and uploading it to a MySQL Database. This includes NaPTAN, NOC, TNDS and BODS data.
 
 The code is separated into Retrievers and Uploaders.
 
 ## Retrievers Description
 
-Python lambdas which are triggered by a CloudWatch Event Cron schedule. The lamdas will retrieve the data from various sources either by HTTP (NaPTAN, NOC) or by FTP (TNDS), unzip it, and then store the data in S3. Due to the size of the TNDS data set, the unzippng process has been split out into its own Lambda which triggers after the data has been retrieved.
+Python lambdas which are triggered by a CloudWatch Event Cron schedule. The lamdas will retrieve the data from various sources either by HTTP (NaPTAN, NOC, BODS) or by FTP (TNDS), unzip it, and then store the data in S3. Due to the size of the TNDS data set, the unzippng process has been split out into its own Lambda which triggers after the data has been retrieved.
 
 ## Uploaders Description
 
@@ -16,7 +16,7 @@ Python lambdas which are triggered after the retrievers have stored the referenc
 
 To import CSV data, the Lambdas run a SQL import script which uses a feature of RDS Aurora to import data directly from S3 into the database, see [here for details](https://docs.aws.amazon.com/AmazonRDS/latest/AuroraUserGuide/AuroraMySQL.Integrating.LoadFromS3.html).
 
-To import the XML TNDS data, the Lambda first parses the XML and then imports the data as necessary into the database.
+To import TransXChange data, the Lambda first parses the XML and then imports the data as necessary into the database.
 
 ## Deploying
