+++ src/retrievers/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "dependencies": {},
   "devDependencies": {
-    "serverless-plugin-aws-alerts": "^1.4.0",
+    "serverless-plugin-aws-alerts": "^1.7.0",
     "serverless-python-requirements": "^5.1.0"
   },
   "scripts": {
